Backend/PlantDiseaseIdentification_JsIntegration/requirements.txt
Django==4.1
djangorestframework==3.13.1
Pillow==9.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
